src/api/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
numpy
pypdf
python-dotenv
bs4

langchain
langchain-core
langchain-community
langchain-ollama
langchain-openai
langchain_chroma